CS2621 Emulator v0.1.0 ComBase

ComBase emulates a distribtued application connected over a potentially asynchronous network.

Functions

append_fuzzers(ctx, fuzzer_list)

Specs

append_fuzzers(
  %ComBase{
    fuzz_chain: pid(),
    registration: term(),
    rev_registration: term(),
    times: term(),
    unfuzzables: term()
  },
  [{atom(), float() | pid()}]
) :: :ok

Add to the list of fuzzers used when messages are received. Note this function must be called before any messages are sent.

broadcast(ctx, msg)

Specs

broadcast(
  %ComBase{
    fuzz_chain: term(),
    registration: pid(),
    rev_registration: pid(),
    times: term(),
    unfuzzables: term()
  },
  any()
) :: boolean()

Send message to all active processes.

get_fuzzers(ctx)

Specs

get_fuzzers(%ComBase{
  fuzz_chain: pid(),
  registration: term(),
  rev_registration: term(),
  times: term(),
  unfuzzables: term()
}) :: [{atom(), float() | pid()}]

Get the current fuzzer list

init()

Specs

init() :: %ComBase{
  fuzz_chain: pid(),
  registration: pid(),
  rev_registration: pid(),
  times: pid(),
  unfuzzables: pid()
}

Initialize the communication base.

list_proc(ctx)

Specs

list_proc(%ComBase{
  fuzz_chain: term(),
  registration: pid(),
  rev_registration: term(),
  times: term(),
  unfuzzables: term()
}) :: [atom()]

Get a list of all registed processes

mark_unfuzzable(ctx)

Specs

mark_unfuzzable(%ComBase{
  fuzz_chain: term(),
  registration: term(),
  rev_registration: term(),
  times: term(),
  unfuzzables: pid()
}) :: :ok

Mark this process as one whose messages (i.e., those sent or received by the process) should not be fuzzed. This is meant as an aid to testing, and should not be used by code not within a test.

now(ctx)

Specs

now(%ComBase{
  fuzz_chain: term(),
  registration: term(),
  rev_registration: term(),
  times: pid(),
  unfuzzables: term()
}) :: number()

Get current time.

send(ctx, proc, msg)

Specs

send(
  %ComBase{
    fuzz_chain: term(),
    registration: pid(),
    rev_registration: pid(),
    times: term(),
    unfuzzables: term()
  },
  atom() | pid(),
  any()
) :: boolean()

Send a message to the process named proc. Message can be anything. ctx should be the context in which the process was created.

set_time(ctx, time)

Specs

set_time(
  %ComBase{
    fuzz_chain: term(),
    registration: term(),
    rev_registration: term(),
    times: pid(),
    unfuzzables: term()
  },
  number()
) :: :ok

Set current time.

spawn(ctx, name, f)

Specs

spawn(
  %ComBase{
    fuzz_chain: term(),
    registration: pid(),
    rev_registration: pid(),
    times: pid(),
    unfuzzables: term()
  },
  atom(),
  (() -> any())
) :: pid()

Spawn a process with supplied name and function.

terminate(ctx)

Specs

terminate(%ComBase{
  fuzz_chain: pid(),
  registration: pid(),
  rev_registration: pid(),
  times: pid(),
  unfuzzables: term()
}) :: :ok

Terminate communication base.

timer(ctx, ms)

Specs

timer(
  %ComBase{
    fuzz_chain: term(),
    registration: pid(),
    rev_registration: pid(),
    times: term(),
    unfuzzables: term()
  },
  non_neg_integer()
) :: reference()

Set a timer for the given number of milliseconds, and send a message with the atom :timer when done.

timer(ctx, ms, atom)

Specs

timer(
  %ComBase{
    fuzz_chain: term(),
    registration: pid(),
    rev_registration: pid(),
    times: term(),
    unfuzzables: term()
  },
  non_neg_integer(),
  atom()
) :: reference()

Set a timer for the given number of milliseconds, and send a message with the atom message when done.

translate_time(ctx, p, time)

Specs

translate_time(
  %ComBase{
    fuzz_chain: term(),
    registration: term(),
    rev_registration: term(),
    times: pid(),
    unfuzzables: term()
  },
  atom(),
  number()
) :: number()

Translate time (gotten from System.monotonic_time()) to time at process p. This should only be used for testing.

whoami(ctx)

Specs

whoami(%ComBase{
  fuzz_chain: term(),
  registration: term(),
  rev_registration: pid(),
  times: term(),
  unfuzzables: term()
}) :: atom() | pid()

Get ID for the current process.