CS2621 Emulator v0.1.0 Emulation

This module is a wrapper around ComBase that simplifies its use in student code. It does so by saving the context as a named Agent, and retrieving it when necessary.

Functions

append_fuzzers(fuzzer_list)

Specs

append_fuzzers([{atom(), float() | pid()}]) :: :ok

Add to the list of fuzzers used when messages are received. Note this function must be called before any messages are sent.

broadcast(msg)

Specs

broadcast(any()) :: boolean()

Send message to all active processes.

cancel_timer(timer)

Specs

cancel_timer(reference()) :: number() | false

Cancel a timer previously set using either timer/1 or timer/2. timer is the reference returned by those calls. Returns the number of milliseconds left on the timer or false if the timer had already expired.

emu_to_micros(time)

Specs

emu_to_micros(number()) :: number()

Convert emulation time units (e.g., from succesive calls to now) into microsecons. This is useful for human readable output.

emu_to_millis(time)

Specs

emu_to_millis(number()) :: number()

Convert emulation time units (e.g., from succesive calls to now) into milliseconds. This is useful for human readable output.

get_fuzzers()

Specs

get_fuzzers() :: [{atom(), float() | pid()}]

Get the current fuzzer list

init()

Specs

init() :: {:ok, pid()}

Setup the emulation

list_proc()

Specs

list_proc() :: [atom()]

Get a list of all registed processes

mark_unfuzzable()

Specs

mark_unfuzzable() :: :ok

Mark this process as one whose messages (i.e., those sent or received by the process) should not be fuzzed. This is meant as an aid to testing, and should not be used by code not within a test.

millis_to_emu(time)

Specs

millis_to_emu(number()) :: number()

Convert milliseconds to emulation time units. This is useful when providing estimates.

now()

Specs

now() :: number()

Get current time.

send(proc, msg)

Specs

send(atom() | pid(), any()) :: boolean()

Send a message to the process named proc. Message can be anything. ctx should be the context in which the process was created.

set_time(time)

Specs

set_time(number()) :: :ok

Set current time.

spawn(name, f)

Specs

spawn(atom(), (() -> any())) :: pid()

Spawn a process with supplied name and function.

terminate()

Teardown emulation

timer(ms)

Specs

timer(non_neg_integer()) :: reference()

Set a timer for ms milliseconds, and send a message to the calling process with the atom :timer when done.

timer(ms, atom)

Specs

timer(non_neg_integer(), atom()) :: reference()

Set a timer for ms milliseconds, and send a message to the calling process with atom when done.

translate_time(p, time)

Specs

translate_time(atom(), number()) :: number()

Translate time (gotten from System.monotonic_time()) to time at process p. This should only be used for testing.

whoami()

Specs

whoami() :: atom() | pid()

Get ID for the current process.