CS2621 Emulator v0.1.0 Fuzzers

Fuzzers is a set of network fuzzing tools.

Functions

build_fuzz_chain(recver, recver_id, sender_id, chain)

Specs

build_fuzz_chain(pid(), pid() | atom(), pid() | atom(), [
  {atom(), float() | pid() | {float(), pid()}}
]) :: pid()

Build a chain of network fuzzing steps.

delay(t)

Specs

delay(number()) :: {:delay, float()}

Fuzzer for delaying messages. Delays are drawn from an exponential distribution with the mean given here.

drop(prob)

Specs

drop(float()) :: {:drop, float()}

Fuzzer for dropping packets. Drops are uniformly random with supplied probability.

drop_record(p, arg)

Specs

drop_record(float(), {:linearize, pid()}) :: {:rec_drop, {float(), pid()}}

Add a fuzzer that drops messages, and records them in the supplied logger.

get_logs(arg)

Specs

get_logs({:linearize, pid()}) :: [{atom() | pid(), atom() | pid(), any()}]

Return a list of message logs from a recording.

log_messages()

Specs

log_messages() :: {:linearize, pid()}

Add a fuzzer that logs messages. Use get_logs/1 to retrieve the messages.